requirements.txt
faker
pyodb
crewai
crewai-tools
PyYAML
